Added a revised CrxConfiguration.h header received from TM.
"As per suggestion, we have modified the CrxConfiguration.h file by declaring couple of public static constants in the Configuration namespace. We prefer to use const rather than public enum, because if we are using enums to define then we have to declare Settings structures (CrxStructGeneralSetting and CrxStructPwvSetting) as enum also. By doing so, it may have impact in other methods and projects. 
-Smarajit

Just to add further clarity, if we declare these numbers as enums, we will also need to change the data type of members of the general and pwv settings structures. This will have larger impact on the code. 
Shriganesh Chavan"



git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk/PN022@183 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97