Patched some fixes from TM:
* DAL 0.1.3.0.3 (/For Atcor/Drop 2/Sprint 3/intermediate/dal 0.1.3.0.3)
-to fix "Data Capture Error Invalid Packet. Do you want to cancel the Capture process?" error during live capture

Known issue:
There is a small issue with GUI which we noticed when theTimeout message is displayed. After the message box is displayed and user clicks NO, If you click on cancel, the same message is displayed again. This does not mean that the timeout event was raised again. The GUI has been fixed but the changes will not be sent with this patch.

* GUI PWVSW-328 (DFCT-IT OnShowStatusEvent failed when called for ErrorSeverity::Exception)
-to fix "Unable to cast object of type 'AtCor.Scor.BusinessLogic.BizPWV' to type 'System.Windows.Forms.IWin32window'"


git-svn-id: https://sfs.atcorau.local/svn/SoftwareRepository/trunk/PN022@352 a8a9b14d-6b92-5b4b-9153-4cd90bb22c97